OCM-3833 | refactor: interactive helpers
