+++ cmd/create/machinepool/machinepool.go
@@ -11,12 +11,12 @@ import (
 	"github.com/briandowns/spinner"
 	diskValidator "github.com/openshift-online/ocm-common/pkg/machinepool/validations"
 	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
-	"github.com/openshift/rosa/pkg/aws"
 	"github.com/openshift/rosa/pkg/helper"
 	mpHelpers "github.com/openshift/rosa/pkg/helper/machinepools"
 	"github.com/openshift/rosa/pkg/helper/versions"
 	"github.com/openshift/rosa/pkg/interactive"
 	"github.com/openshift/rosa/pkg/interactive/confirm"
+	interactiveSgs "github.com/openshift/rosa/pkg/interactive/securitygroups"
 	"github.com/openshift/rosa/pkg/ocm"
 	"github.com/openshift/rosa/pkg/output"
 	"github.com/openshift/rosa/pkg/rosa"
@@ -291,30 +291,7 @@ func addMachinePool(cmd *cobra.Command, clusterKey string, cluster *cmv1.Cluster
 			r.Reporter.Warnf("Unexpected situation a VPC ID should have been selected based on chosen subnets")
 			os.Exit(1)
 		}
-		possibleSgs, err := r.AWSClient.GetSecurityGroupIds(vpcId)
-		if err != nil {
-			r.Reporter.Errorf("There was a problem retrieving security groups for VPC '%s': %v", vpcId, err)
-			os.Exit(1)
-		}
-		if len(possibleSgs) > 0 {
-			options := []string{}
-			for _, sg := range possibleSgs {
-				options = append(options, aws.SetSecurityGroupOption(sg))
-			}
-			securityGroupIds, err = interactive.GetMultipleOptions(interactive.Input{
-				Question: "Additional Security Group IDs",
-				Help:     cmd.Flags().Lookup(securityGroupIdsFlag).Usage,
-				Required: false,
-				Options:  options,
-			})
-			if err != nil {
-				r.Reporter.Errorf("Expected valid Security Group IDs: %s", err)
-				os.Exit(1)
-			}
-			for i, sg := range securityGroupIds {
-				securityGroupIds[i] = aws.ParseOption(sg)
-			}
-		}
+		securityGroupIds = interactiveSgs.GetSecurityGroupIds(r, cmd, vpcId, securityGroupIdsFlag)
 	}
 	for i, sg := range securityGroupIds {
 		securityGroupIds[i] = strings.TrimSpace(sg)
